+++ src/lib/Backend/Py/InsightsGraph.py
@@ -7,7 +7,6 @@ if len(sys.argv) < 5:
 
 labels = sys.argv[1]
 labels = labels.split(',')
-# text == ['a', 'b', 'c']
 
 joins = sys.argv[2]
 joins = joins.split(',')
@@ -17,6 +16,7 @@ plt.rcParams['text.color'] = '#99AAB5'
 plt.rcParams['axes.labelcolor'] = '#99AAB5'
 plt.rcParams['xtick.color'] = '#99AAB5'
 plt.rcParams['ytick.color'] = '#99AAB5'
+
 # how tall Y axis should be
 y = np.arange(len(labels))
 
@@ -30,4 +30,4 @@ ax = plt.gca()
 ax.set_facecolor('#2C2F33')
 plt.title('Discord Guild Insights')
 
-plt.savefig(sys.argv[4] + sys.argv[3] + '.png')
+plt.savefig(sys.argv[4] + sys.argv[3] + '.jpg', format='jpg', optimize=True)